Struct temporal_networks::interval::Interval

pub struct Interval(pub f64, pub f64);

An interval represents a context-agnostic inclusive [lower, upper] time range. While Interval may be accessible from JS, the Rust implementation includes additional operator overloads for simplified arithmetic.

JS-specific

Use .toJSON() if you want to convert an Interval to an array of numbers

Using indexing to get the lower and upper bounds is also an option, eg. interval[0] === lower && interval[1] === upper

use Number.MAX_VALUE and -Number.MAX_VALUE to represent infinity and -infinity respectively

Examples

Interval arithmetic in Rust.

use temporal_networks::interval::Interval;

let interval1 = Interval::new(0., 10.);
let interval2 = Interval::new(5., 16.);

let summed_interval = Interval::new(5., 26.);
assert_eq!(interval1 + interval2, summed_interval);

let diff_interval = Interval::new(-5., 16.);
assert_eq!(interval2 - interval1, diff_interval);

let unioned_interval = Interval::new(5., 10.);
assert_eq!(interval1 & interval2, unioned_interval);

Methods

impl Interval

pub fn new(lower: f64, upper: f64) -> Interval

Create a new Interval

pub fn from_vec(other: Vec<f64>) -> Interval

Get an interval from a vector

pub fn to_json(&self) -> JsValue

Convert the interval to JSON [lower, upper]

pub fn lower(&self) -> f64

The lower bound of the range

pub fn upper(&self) -> f64

The upper bound of the range

pub fn contains(&self, v: f64) -> bool

Whether or not a point in time falls within a range

pub fn is_valid(&self) -> bool

A check that ensures the lower bound is less than the upper bound

pub fn converged(&self) -> bool

Whether or not the interval has converged to a time