Module temporal_networks::plan

Plan

Defines an API designed to be exported to WASM that can perform time math without requiring the user to understand the underlying data structures or algorithms.

Nomenclature (and some types)

Plan: a set of temporal constraints describing a set of actions to occur. Our implementation of Plan can currently handle actions that occur in series or parallel (or any mix thereof)

Event: a moment in time in the Plan

Step: A pair of start and end Events

Interval: A span of time represented in [lower, upper] range

Duration: An interval in the context of a Step, ie, the interval between the start and end events. In English, a Step with a [lower, upper] duration is "an action that will take between lower and upper units of time to complete".

Structs

Plan
A Plan orchestrates events and the timing constraints between them. It allows for querying arbitrary timing information with knowledge of the underlying data structure.

Step
A Step represents a logical action that occurs over a period of time. It implicitly has start and end events, which are used by Plan