Struct temporal_networks::plan::Plan

pub struct Plan { /* fields omitted */ }

A Plan orchestrates events and the timing constraints between them. It allows for querying arbitrary timing information with knowledge of the underlying data structure.

Example

Creating a Plan and adding Steps with constraints in Rust

use temporal_networks::plan::Plan;
use temporal_networks::interval::Interval;

// create a plan
let mut plan = Plan::new();

// add a step to the plan that takes between 6 and 17 time units to complete
let step1 = plan.add_step("example".to_string(), Some(vec![6., 17.]));

// add another step and a constraint that the second occurs after the first
let step2 = plan.add_step("another example".to_string(), Some(vec![8., 29.]));
plan.add_constraint(step1.end(), step2.start(), None);

// find the [lower, upper] interval between the start of the plan and the start of the second step
let root = plan.root().unwrap();
let result = plan.interval(root, step2.start()).unwrap();

// you may notice the interval between the start of the plan and the second step is just the duration of the first step!
assert_eq!(result, Interval::new(6., 17.));

Methods

impl Plan

pub fn new() -> Plan

pub fn root(&mut self) -> Option<i32>

Get the first event in the plan. Found implicitly based on the current constraints

pub fn create_event(&mut self, identifier: String) -> i32

Low-level API for creating nodes in the graph. Advanced use only. If you can't explain why you should use this over addStep, use addStep instead

pub fn add_step(
&mut self,
identifier: String,
duration: Option<Vec<f64>>
) -> Step

Create a new step and add it to this plan. The identifier is recommended but not required to be unique (being unique may become a requirement in the future)

pub fn get_duration(&self, s: &Step) -> Interval

Get the controllable duration of a step

pub fn compile(&mut self) -> Result<(), JsValue>

Compile the plan into a dispatchable form. A dispatchable form is required to query the plan for almost any scheduling information. This method is called implicitly when you attempt to query the plan when the dispatchable graph is not up-to-date. However, you can proactively call compile at a time that is computationally convenient for your application to avoid paying the performance penalty when querying the plan

pub fn complete_step()

pub fn interval(
&mut self,
source: i32,
target: i32
) -> Result<Interval, JsValue>

Get the interval between two events

pub fn event_distance(
&mut self,
source: i32,
target: i32
) -> Result<JsValue, JsValue>

Low-level API to get the directional distance between two events. Advanced use only. If you can't explain why you should use this over interval, use interval instead

pub fn update_duration()

pub fn add_constraint(
&mut self,
source: i32,
target: i32,
interval: Option<Vec<f64>>
) -> Result<(), JsValue>

Add a constraint between the start or end of two events. Errs if either source or target is not already in the plan. Defaults to a [0, 0] interval between events

pub fn remove_constrainst()

pub fn remove_constrainsts()