Module temporal_networks::schedule

Schedule

Defines an API designed to be exported to WASM that can perform time math without requiring the user to understand the underlying data structures or algorithms.

Nomenclature (and some types)

We're drawing terminology from existing planners and the literature on temporal plans. Eg. see [1] pg 519 (also in docs/references/).

Schedule: a set of temporal constraints describing a set of Episodes to occur. Our implementation of Schedule can currently handle Episodes that occur in series or parallel (or any mix thereof)

Event: a moment in time in the Schedule

Episode: A pair of start and end Events

Interval: A span of time represented in [lower, upper] range

Duration: An interval in the context of a Episode, ie, the interval between the start and end events. In English, an Episode with a [lower, upper] duration is "an episode that will take between lower and upper units of time to complete".

[1] Ono, M., Williams, B. C., & Blackmore, L. (2013). Probabilistic planning for continuous dynamic systems under bounded risk. Journal of Artificial Intelligence Research, 46, 511–577. https://doi.org/10.1613/jair.3893

Structs

Episode
An Episode represents a logical action that occurs over a period of time. It implicitly has start and end events, which are used by Schedule

Schedule
A Schedule orchestrates events and the timing constraints between them. It allows for querying arbitrary timing information with knowledge of the underlying data structure.

Type Definitions

EventID
An ID representing an event in the Schedule